# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "nanoblogging platform as PWA built on go-app framework",
        "title": "litter-go",
        "termsOfService": "https://littr.n0p.cz/tos",
        "contact": {
            "name": "API Support",
            "url": "https://littr.n0p.cz/docs/",
            "email": "littr@n0p.cz"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/krustowski/litter-go/blob/master/LICENSE"
        },
        "version": "0.35.0"
    },
    "host": "littr.n0p.cz",
    "basePath": "/api/v1",
    "paths": {
...
        "/polls/": {
            "get": {
                "description": "get a list of polls",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Get a list of polls",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "add new poll",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Add new poll",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/users/": {
            "get": {
                "description": "get a list of users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get a list of users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "add new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Add new user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "auth tokens (JWT)",
                    "type": "string"
                },
                "auth_granted": {
                    "type": "boolean"
                },
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Device"
                    }
                },
                "flow_records": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "flow_stats": {
                    "description": "very stats properties",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "polls": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.Poll"
                    }
                },
                "posts": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.Post"
                    }
                },
                "public_key": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "subscription": {
                    "type": "object",
                    "properties": {
                        "mentions": {
                            "type": "boolean"
                        },
                        "replies": {
                            "type": "boolean"
                        }
                    }
                },
                "user_stats": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.UserStat"
                    }
                },
                "users": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.Device": {
            "type": "object",
            "properties": {
                "subscription": {
                    "description": "The very subscription struct/details.\nSubscription app.NotificationSubscription `json:\"subscription\"`",
                    "allOf": [
                        {
                            "$ref": "#/definitions/webpush.Subscription"
                        }
                    ]
                },
                "tags": {
                    "description": "List of labels for such device.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time_created": {
                    "description": "Timestamp of the subscription creation.",
                    "type": "string"
                },
                "time_last_used": {
                    "description": "Timestamp of the last notification sent through this device.",
                    "type": "string"
                },
                "uuid": {
                    "description": "Unique identification of the app on the current device.\nhttps://go-app.dev/reference#Context",
                    "type": "string"
                }
            }
        },
        "models.Poll": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Author is the back key to the user originally posting that poll.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is an unique poll's identifier.",
                    "type": "string"
                },
                "option_one": {
                    "description": "OptionOne is the answer numero uno.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PollOption"
                        }
                    ]
                },
                "option_three": {
                    "description": "OptionThree is the answer numero tres.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PollOption"
                        }
                    ]
                },
                "option_two": {
                    "description": "OptionTwo is the answer numero dos.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.PollOption"
                        }
                    ]
                },
                "question": {
                    "description": "Question is to describe the main purpose of such poll.",
                    "type": "string"
                },
                "reaction_count": {
                    "description": "ReactionCount counts the number of item's reactions.",
                    "type": "integer"
                },
                "timestamp": {
                    "description": "Timestamp is an UNIX timestamp indication the poll's creation time; should be identical to the upstream post's Timestamp.",
                    "type": "string"
                },
                "voted_list": {
                    "description": "VodeList is the list of user nicknames voted on such poll already.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.PollOption": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content describes the very content of such poll's option/answer.",
                    "type": "string"
                },
                "counter": {
                    "description": "Counter hold a number of votes being committed to such option.",
                    "type": "integer"
                }
            }
        },
        "models.Post": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Content contains the very post's data to be shown as a text typed in by the author when created.",
                    "type": "string"
                },
                "data": {
                    "description": "Data is a helper field for the actual figure upload.",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "figure": {
                    "description": "Figure hold the filename of the uploaded figure to post with some provided text.",
                    "type": "string"
                },
                "id": {
                    "description": "ID is an unique post's identificator.",
                    "type": "string"
                },
                "nickname": {
                    "description": "Nickname is a name of the post's author's name.",
                    "type": "string"
                },
                "poll_id": {
                    "description": "PollID is an identification of the Poll structure/object.",
                    "type": "string"
                },
                "reaction_count": {
                    "description": "ReactionCount counts the number of item's reactions.",
                    "type": "integer"
                },
                "reply_count": {
                    "description": "ReplyCount hold the count of replies for such post.",
                    "type": "integer"
                },
                "reply_to": {
                    "description": "ReplyTo is a reference key to another post, that is being replied to.",
                    "type": "integer"
                },
                "reply_to_id": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "Timestamp is an UNIX timestamp, indicates the creation time.",
                    "type": "string"
                },
                "type": {
                    "description": "Type describes the post's type --- post, poll, reply, img.",
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "id",
                "nickname"
            ],
            "properties": {
                "about": {
                    "description": "About is a description string of such user.",
                    "type": "string"
                },
                "active": {
                    "description": "Active boolean indicates an activated user's account.",
                    "type": "boolean"
                },
                "app_bg_mode": {
                    "description": "AppBgMode string defines the colour mode of the app's background (light vs dark).",
                    "type": "string",
                    "default": "dark"
                },
                "avatar_url": {
                    "description": "AvatarURL is an URL to the user's custom profile picture.",
                    "type": "string"
                },
                "color": {
                    "description": "Color is the user's UI color scheme.",
                    "type": "string",
                    "default": "#000000"
                },
                "email": {
                    "description": "Email is a primary user's e-mail address.",
                    "type": "string"
                },
                "flow_list": {
                    "description": "FlowList is a string map of users, which posts should be added to one's flow page.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "flow_toggle": {
                    "description": "FlowToggle is a single implementation of FlowList.",
                    "type": "string"
                },
                "full_name": {
                    "description": "FullName is the \"genuine\" name of such user.",
                    "type": "string"
                },
                "gdpr": {
                    "description": "GDPR consent, set to true because it is noted on the registration page so. No user data should\nbe saved if the boolean is false.",
                    "type": "boolean"
                },
                "id": {
                    "description": "ID is an unique identifier.",
                    "type": "string"
                },
                "last_active_time": {
                    "description": "LastLoginTime is an UNIX timestamp of the last action performed by such user.",
                    "type": "string"
                },
                "last_login_time": {
                    "description": "LastLoginTime is an UNIX timestamp of the last user's successful log-in.",
                    "type": "string"
                },
                "nickname": {
                    "description": "Nickname is a login name of such user.",
                    "type": "string"
                },
                "passphrase": {
                    "description": "Passphrase is a hashed pass phrase string (binary form).",
                    "type": "string"
                },
                "passphrase_hex": {
                    "description": "PassphraseHex is a hashed pass phrase string (hexadecimal alphanumberic form).",
                    "type": "string"
                },
                "private": {
                    "description": "Private boolean indicates a private user's account.",
                    "type": "boolean"
                },
                "registered_time": {
                    "description": "RegisteredTime is an UNIX timestamp of the user's registeration.",
                    "type": "string"
                },
                "request_list": {
                    "description": "RequestList is a map of account requested to add this user to their flow --- used with the Private property.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "shade_list": {
                    "description": "ShadeList is a map of account/users to be shaded (soft-blocked) from following.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "boolean"
                    }
                },
                "tags": {
                    "description": "Tags is an array of possible roles and other various attributes assigned to such user.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "web": {
                    "description": "Web is user's personal homepage.",
                    "type": "string"
                }
            }
        },
        "models.UserStat": {
            "type": "object",
            "properties": {
                "flowerCount": {
                    "description": "FlowerCount is basically a number of followers.",
                    "type": "integer"
                },
                "postCount": {
                    "description": "PostCount is a number of posts of such user.",
                    "type": "integer"
                },
                "reactionCount": {
                    "description": "ReactionCount tells the number of interactions (stars given).",
                    "type": "integer"
                },
                "searched": {
                    "description": "Searched is a special boolean used by the search engine to mark who is to be shown in search results.",
                    "type": "boolean"
                },
                "shadeCount": {
                    "description": "ShadeCount is basically a number of blockers.",
                    "type": "integer"
                }
            }
        },
        "webpush.Keys": {
            "type": "object",
            "properties": {
                "auth": {
                    "type": "string"
                },
                "p256dh": {
                    "type": "string"
                }
            }
        },
        "webpush.Subscription": {
            "type": "object",
            "properties": {
                "endpoint": {
                    "type": "string"
                },
                "keys": {
                    "$ref": "#/definitions/webpush.Keys"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
    "security": [
        {
            "": [
                ""
            ]
        }
    ],
...
}